Revert "Implemented Day 13 Part 1 solution."
This reverts commit 6ae7c5d83cb2aadb63c92c9a0b89ecb01abdaefe.

+++ AdventOfCode2024/Problems/Day12.cs
@@ -12,8 +12,8 @@ namespace AdventOfCode2024.Problems
 
         #region Fields
         string _inputPath = @"PASTE PATH HERE"; 
-        long _firstResult = 0;
-        long _secondResult = 0;
+        int _firstResult = 0;
+        ulong _secondResult = 0;
         string[] _gardenPlot = [];
         char[] _distinctPlotValues = [];
 
@@ -35,7 +35,7 @@ namespace AdventOfCode2024.Problems
         }
 
 
-        long FirstResult
+        int FirstResult
         {
             get => _firstResult;
             set
@@ -46,7 +46,7 @@ namespace AdventOfCode2024.Problems
                 }
             }
         }
-        long SecondResult
+        ulong SecondResult
         {
             get => _secondResult;
             set
@@ -99,8 +99,8 @@ namespace AdventOfCode2024.Problems
         public Day12()
         {
             InitialiseProblem();
-            FirstResult = SolveFirstProblem<long>();
-            SecondResult = SolveSecondProblem<long>();
+            FirstResult = SolveFirstProblem<int>();
+            SecondResult = SolveSecondProblem<ulong>();
             OutputSolution();
         }
         #endregion
